Update Heading Size
